LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/navigating-across-documents/014-expected.txt


FAIL  Link with javascript onclick form submission script order  assert_array_equals: expected property 0 to be "after script" but got "submit" (expected array ["after script", "submit"] got ["submit", "after script"])

Test
